+++ transport/grpc/headers.go
@@ -86,6 +86,11 @@ const (
 
 	baseContentType   = "application/grpc"
 	contentTypeHeader = "content-type"
+
+	// grpcAcceptEncodingHeader is the header name that keeps the list of available compressors,
+	// applicable for both request and response headers. Usually consumed by grpc, but the
+	// value is kept in metadata and available in application code.
+	grpcAcceptEncodingHeader = "grpc-accept-encoding"
 )
 
 // TODO: there are way too many repeat calls to strings.ToLower
@@ -128,7 +133,12 @@ func metadataToTransportRequest(md metadata.MD) (*transport.Request, error) {
 		case 1:
 			value = values[0]
 		default:
-			return nil, yarpcerrors.InvalidArgumentErrorf("header has more than one value: %s:%v", header, values)
+			// TODO (https://github.com/yarpc/yarpc-go/issues/2265) don't pass this values to application code
+			if header == grpcAcceptEncodingHeader {
+				value = values[0]
+			} else {
+				return nil, yarpcerrors.InvalidArgumentErrorf("header has more than one value: %s:%v", header, values)
+			}
 		}
 		header = transport.CanonicalizeHeaderKey(header)
 		switch header {
@@ -213,7 +223,12 @@ func getApplicationHeaders(md metadata.MD) (transport.Headers, error) {
 		case 1:
 			value = values[0]
 		default:
-			return headers, yarpcerrors.InvalidArgumentErrorf("header has more than one value: %s:%v", header, values)
+			// TODO (https://github.com/yarpc/yarpc-go/issues/2265) don't pass this values to application code
+			if header == grpcAcceptEncodingHeader {
+				value = values[0]
+			} else {
+				return headers, yarpcerrors.InvalidArgumentErrorf("header has more than one value: %s:%v", header, values)
+			}
 		}
 		headers = headers.With(header, value)
 	}